pyproject.toml
[tool.poetry]
name = "msu_test"
version = "0.1.0"
description = ""
authors = ["Alexey <work.alexeykudimov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
Pillow = "^8.4.0"
psycopg2 = "^2.9.2"
drf-yasg = "^1.20.0"
djangorestframework-simplejwt = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
